requirements.txt
streamlit>=1.38.0
pandas>=2.0.0
yfinance>=0.2.40
plotly>=5.22.0
openai>=1.40.0
google-generativeai>=0.8.0
